Proper output using an output buffer
Output so far was also done using STDOUT directly. Now we assume that STDOUT
can be not ready at times and use stream_select to determine if it is ready too.
However, this leads to the next problem: STDOUT is basically ready all the time.
So we should only check if there is actually something to write.

There is a conceptional bug - how to add the prompt "> " which should not have
a new line? You'll see.