Added a longer running task to timeshare.php
Added command sha1 <directory>, which uses DirectorySHA to calculate an SHA1 sum
of each and every file in and below a directory.
This is meant to show that timeshare.php is still able to respond to user input,
even if multiple processes are running.

It also shows a flaw: if DirectorySHA1 processes a larger file, it of course
blocks.
Implementing my own SHA1-function which is timeshareable would not solve the
problem yet - it would be necessary that DirectorySHA gives its timeslot to
the SHA1-function and then have DirectorySHA proceed after SHA1 has finished.