Turned Timeshare into a regular class again and have it implement Timeshared
With Timeshare being an implementation of Timeshared itself, we can easily
implement tree like structures of subprocesses dependent on another process.

This is not well thought out yet. If the topmost implementation had to processes
A and B, but A would be an instance of Timeshared, but B not, B would get 50%
of the runtime, while all processes of A would have to accomodate with the
remaining 50%.
However, if this would go beyond a mere proof of concept, scheduling would
become a question in itself. As an example, a process waiting for input
basically only have to evaluate if fread is false or not most of the time. It
would not make sense to give this process the same 50% as, say, SHA1.