DirectoryLinear now implements some sort of queue/concurrency for FileHandlers
DirectoryLinear now uses Timeshared to run handlers as returned by the
FileHandlerFactory. Up to five file handlers will be processed along
with DirectoryLinear itself; once five handlers are reached,
DirectoryLinear will stop to traverse the directory structure.

Now this does not make the handlers run faster, of course not...
...but from a certain perspective, it does, as can be demonstrated with
the SHAFile example. Once a large file is hit, smaller files continue to
be processed. While this will slow down the processing of the larger file,
it still allows for many smaller files to be processed.

Five was just arbitrary chosen to test the concept.