Properly implemented Timeshared::terminate()
When terminate is called on an implementation of Timeshared, the application is
supposed to properly finish whatever it is doing and then have loop() return
false. Once the process queue is empty, Timeshare will end as well.

There probably needs to be some kind of timeout mechanism for processes that
do not terminate correctly or in time.