Made changes to reimplement "halt" again.
Having the "Server side shutdown" message in onDisconnect/calling onDisconnect
is not correct, as this should be called if the client disconnects. Works for
now.

Shutting down is not as trivial as it seems if you want to make sure that the
server does not go offline before sending out the message.