Using separate Timeshare instance for Clients.
Having ServerProcess & each client stream in one instance of Timeshare means
that it is difficult to implement 'halt', as ServerProcess::terminate must
not call Timeshare::terminate (which leads to an endless loop of both calling
each other). The drawback however is that the main Timeshare is skewed, calling
ServerProcess and the client Timeshare 50/50.