Command 'count' which counts from 1 to 15, every second.
This is one possible solution how a non-interactive process is started: Main
delegates to Counter. When Counter itself is done, Counter::loop returns false
and control is given back to Main.
Currently, Main::terminate immediately stops the delegate and returns control
to Main, which then adds the server side shutdown message.

This is no problem for Counter, but it would be a problem for, say, a file
transfer - as the client would need to go into a different mode too, it would
not react to quit and need another way of getting notified that file transfer
has been aborted prematurely.