Added stream_select to Stream::loop, StreamBinary::write throws Exception.
An non blocking stream may just drop data if it is not ready to be written to,
so if there is data to write, it is only written if the stream is ready to be
written.
Same goes for read, data is only read if there is data to be read.