Implemented upload of files
'put filename' can be used to upload files. Mechanism is a little sketchy so
far, basically the idea is that the server receives 'put' and uses a delegate
to switch into receive mode while the client uses a delegate to switch into
the send mode.

Right now, it is still verbose for testing and the implementation is not very
elegant, but it is working fine.