Changed ServerMain into Singleton
I try not to use Singletons, but this is an acceptable use case in that it
allows to halt the server from other parts of the programm without having
to dependency inject it again and again. And as ServerMain is only used in
one instance, using a Singleton does not hurt.