Refactored StreamListener to StreamHandler.
Classes such as Main and Counter are not observers; in the end, their main
task is to handle/provide stream data. One could see a Strategy Pattern at
play here, but somehow this does not really fit namewise, so I stuck with
StreamHandler. The two observer methods onTerminate/onDisconnect are
still part of StreamHandler; one could use a distinct, additional observer.

I refactored the main command routines as well as the Counter class. File
transfer is not yet refactored.